requirements.txt
discord==1.0.1
numpy==1.19.2
pymysql==0.10.1
pytz==2020.1
python-dateutil==2.4.1
validator_collection==1.5.0
discord-py-interactions==4.1.0
